tried using double, got nan elbo in 70 iters. Reverting back to float

+++ cytopy/util/simdata.py
@@ -12,7 +12,7 @@ def simdata(N=[300, 100, 200], J=25, a_W=[200., 500., 200., 100.], L0=5, L1=3, s
     I = len(N)
     data = {'y': [], 'm': []}
     K = len(a_W)
-    a_W = torch.tensor(a_W).double()
+    a_W = torch.tensor(a_W).float()
 
     if alpha is None:
         alpha = K
@@ -37,8 +37,8 @@ def simdata(N=[300, 100, 200], J=25, a_W=[200., 500., 200., 100.], L0=5, L1=3, s
     if sig is None:
         sig = torch.ones((I, )) * 1.0
 
-    mu0 = -(torch.arange(L0) + 3 * sig.max()).double()
-    mu1 = (torch.arange(L1) + 3 * sig.max()).double()
+    mu0 = -(torch.arange(L0) + 3 * sig.max()).float()
+    mu1 = (torch.arange(L1) + 3 * sig.max()).float()
 
     params = {'W': W, 'v': v, 'eta0': eta0, 'eta1': eta1,
               'mu0': mu0, 'mu1': mu1, 'sig': sig, 'Z': Z}
